Inizio Servlet inizio tirocinio senza controlli
Inserimento della servlet

+++ src/model/Tirocinio.java
@@ -15,7 +15,7 @@ public class Tirocinio {
     private String progettoFormativo;
 	private String competenzeAcquisire;
 	private String descrizioneEnte;
-	private int matricola;
+	private long matricola;
 	private String partitaIva;
 	
 	/** Costruttore vuoto. **/
@@ -41,7 +41,8 @@ public class Tirocinio {
      * @param partitaIva dell' 'EnteConvenzionato' associato
      * 
      */
-    public Tirocinio(int codTirocinio, String dataInizioTirocinio, short cfuPrevisti, String competenze, String competenzeAcquisire, String attivitaPreviste, String svolgimentoTirocinio, String statoTirocinio, String progettoFormativo, String descrizioneEnte, int matricola, String partitaIva) {
+    public Tirocinio(int codTirocinio, String dataInizioTirocinio, short cfuPrevisti, String competenze, String competenzeAcquisire, String attivitaPreviste, String svolgimentoTirocinio, String statoTirocinio, String progettoFormativo, String descrizioneEnte, long
+        matricola, String partitaIva) {
         this.codTirocinio = codTirocinio;
         this.dataInizioTirocinio = dataInizioTirocinio;
         this.cfuPrevisti = cfuPrevisti;
@@ -136,11 +137,11 @@ public class Tirocinio {
 		this.descrizioneEnte = descrizioneEnte;
 	}
 
-	public int getMatricola() {
+	public long getMatricola() {
 		return matricola;
 	}
 
-	public void setMatricola(int matricola) {
+	public void setMatricola(long matricola) {
 		this.matricola = matricola;
 	}
